Move event recorder wrapper to events.go
