+++ operator/pkg/reconciliation/context.go
@@ -5,8 +5,7 @@ package reconciliation
 
 import (
 	"context"
-	"fmt"
-	
+
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -17,6 +16,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/datastax/cass-operator/operator/pkg/httphelper"
+	"github.com/datastax/cass-operator/operator/pkg/events"
 
 	api "github.com/datastax/cass-operator/operator/pkg/apis/cassandra/v1beta1"
 )
@@ -42,27 +42,6 @@ type ReconciliationContext struct {
 	clusterPods            []*corev1.Pod
 }
 
-type LoggingEventRecorder struct {
-	record.EventRecorder
-	ReqLogger logr.Logger
-}
-
-func (r *LoggingEventRecorder) Event(object runtime.Object, eventtype, reason, message string) {
-	r.ReqLogger.Info(message, "reason", reason, "eventType", eventtype)
-	r.EventRecorder.Event(object, eventtype, reason, message)
-}
-
-func (r *LoggingEventRecorder) Eventf(object runtime.Object, eventtype, reason, messageFmt string, args ...interface{}) {
-	r.ReqLogger.Info(fmt.Sprintf(messageFmt, args...), "reason", reason, "eventType", eventtype)
-	r.EventRecorder.Eventf(object, eventtype, reason, messageFmt, args...)
-}
-
-func (r *LoggingEventRecorder) AnnotatedEventf(object runtime.Object, annotations map[string]string, eventtype, reason, messageFmt string, args ...interface{}) {
-	r.ReqLogger.Info(fmt.Sprintf(messageFmt, args...), "reason", reason, "eventType", eventtype)
-	r.EventRecorder.AnnotatedEventf(object, annotations, eventtype, reason, messageFmt, args...)
-}
-
-
 // CreateReconciliationContext gathers all information needed for computeReconciliationActions into a struct.
 func CreateReconciliationContext(
 	req *reconcile.Request,
@@ -75,7 +54,7 @@ func CreateReconciliationContext(
 	rc.Request = req
 	rc.Client = cli
 	rc.Scheme = scheme
-	rc.Recorder = &LoggingEventRecorder{EventRecorder: rec, ReqLogger: reqLogger,}
+	rc.Recorder = &events.LoggingEventRecorder{EventRecorder: rec, ReqLogger: reqLogger,}
 	rc.ReqLogger = reqLogger
 	rc.Ctx = context.Background()
